updated custom extractor
