+++ thingsvision/core/extraction/__init__.py
@@ -1,3 +1,8 @@
-from .helpers import (center_features, create_custom_extractor,
-                      create_model_extractor, get_extractor,
-                      get_extractor_from_model, normalize_features)
+from .helpers import (
+    center_features,
+    create_custom_extractor,
+    create_model_extractor,
+    get_extractor,
+    get_extractor_from_model,
+    normalize_features,
+)
