+++ thingsvision/core/extraction/helpers.py
@@ -7,10 +7,13 @@ import thingsvision.custom_models.cornet as cornet
 import torch
 from torchtyping import TensorType
 
-from .base import BaseExtractor
-from .extractor import (KerasExtractor, SSLExtractor, TimmExtractor,
-                        TorchvisionExtractor)
-from .mixin import PyTorchMixin, TensorFlowMixin
+from .source_extractors import (
+    KerasExtractor,
+    SSLExtractor,
+    TimmExtractor,
+    TorchvisionExtractor,
+)
+from .main_extractors import PyTorchExtractor, TensorFlowExtractor
 
 Array = np.ndarray
 AxisError = np.AxisError
@@ -43,9 +46,9 @@ def create_custom_extractor(
             f"\nCould not find {model_name} among custom models.\nChoose a different model.\n"
         )
 
-    backend_mixin = PyTorchMixin if backend == "pt" else TensorFlowMixin
+    Extractor = PyTorchExtractor if backend == "pt" else TensorFlowExtractor
 
-    class CustomExtractor(BaseExtractor, backend_mixin):
+    class CustomExtractor(Extractor):
         def __init__(self, *args, **kwargs) -> None:
             super().__init__(*args, **kwargs)
 
@@ -140,9 +143,9 @@ def create_model_extractor(
     extractor: Any
         The custom extractor class.
     """
-    backend_mixin = PyTorchMixin if backend == "pt" else TensorFlowMixin
+    Extractor = PyTorchExtractor if backend == "pt" else TensorFlowExtractor
 
-    class ModelExtractor(BaseExtractor, backend_mixin):
+    class ModelExtractor(Extractor):
         def __init__(self, *args, **kwargs):
             super().__init__(*args, **kwargs)
